+++ R/summariseConceptIdCounts.R
@@ -7,7 +7,7 @@
 #' @param countBy Either "record" for record-level counts or "person" for
 #' person-level counts
 #' @param year deprecated
-#' @param interval Time interval to stratify by. It can either be "years", "quarters", "months" or "overall".
+#' @inheritParams interval
 #' @param sex TRUE or FALSE. If TRUE code use will be summarised by sex.
 #' @param ageGroup A list of ageGroup vectors of length two. Code use will be
 #' thus summarised by age groups.
@@ -47,23 +47,20 @@ summariseConceptIdCounts <- function(cdm,
 
     lifecycle::deprecate_warn("0.2.3", "summariseConceptIdCounts(year)", "summariseConceptIdCounts(interval = 'years')")
 
-    if (year & interval == "overall") {
+    if (isTRUE(year) & missing(interval)) {
 
       interval = "years"
-      cli::cli("interval argument set to 'years'")
+      cli::cli_inform("interval argument set to 'years'")
 
-      } else if (year & interval != "overall" ){
+      } else if (isTRUE(year) & !missing(interval)){
 
-       cli::cli("year argument will be ignored")
+       cli::cli_inform("year argument will be ignored")
     }
   }
 
   # initial checks
   cdm <- omopgenerics::validateCdmArgument(cdm)
   checkCountBy(countBy)
-
-  #should i leave the check for year argument?
-
   omopgenerics::assertChoice(interval, c("overall", "years", "quarters", "months"), length = 1)
   omopgenerics::assertLogical(sex, length = 1)
   omopgenerics::assertChoice(omopTableName, choices = omopgenerics::omopTables(), unique = TRUE)