+++ NEWS.md
@@ -1,3 +1,4 @@
 # OmopSketch 0.2.2.900
 
-- Account for int64 in summariseInObservation by @cecicampanile #311
+-   Account for int64 in summariseInObservation by @cecicampanile #311
+-   Interval argument in summariseMissingData and summariseConceptIdCounts, year argument deprecated by @cecicampanile #317